feat: Display and locate distressed users
Introduces functionality to display alerts for other users triggering SOS and allows the current user to quickly navigate to their location.

The SOS sound alert now triggers for both the current user and any other distressed users on the map. A new `SOSAlertOverlay` component is introduced to visually notify users of distress calls, with an option to navigate to the SOS location. The `TacticalOverlay` has been updated to conditionally render this alert.